# src/controller/registry-org.controller/index.js
const express = require('express')
const router = express.Router()
const mw = require('../../middleware/middleware')
const errorMsgs = require('../../middleware/errorMessages')
const { body, param, query } = require('express-validator')
const controller = require('./registry-org.controller')
const { parseGetParams, parsePostParams, parseDeleteParams, parseError, isOrgRole } = require('./registry-org.middleware')
const { toUpperCaseArray, isFlatStringArray } = require('../../middleware/middleware')
const getConstants = require('../../constants').getConstants
const CONSTANTS = getConstants()

router.get('/registryOrg',
  /*
  #swagger.tags = ['Registry Organization']
  #swagger.operationId = 'getAllRegistryOrgs'
  #swagger.summary = "Retrieves information about all registry organizations (accessible to Secretariat only)"
  #swagger.description = "
        <h2>Access Control</h2>
        <p>Only users with Secretariat role can access this endpoint</p>
        <h2>Expected Behavior</h2>
        <p><b>Secretariat:</b> Retrieves a list of all registry organizations</p>
  #swagger.parameters['$ref'] = [
    '#/components/parameters/pageQuery',
    '#/components/parameters/apiEntityHeader',
    '#/components/parameters/apiUserHeader',
    '#/components/parameters/apiSecretHeader'
  ]
  #swagger.responses[200] = {
    description: 'A list of all registry organizations, along with pagination fields if results span multiple pages of data',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/registry-org/get-registry-org-response.json' }
      }
    }
  }
  #swagger.responses[400] = {
    description: 'Bad Request',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/bad-request.json' }
      }
    }
  }
  #swagger.responses[401] = {
    description: 'Not Authenticated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[403] = {
    description: 'Forbidden',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[500] = {
    description: 'Internal Server Error',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  */
  mw.validateUser(true),
  mw.onlySecretariat(true),
  query().custom((query) => { return mw.validateQueryParameterNames(query, ['page']) }),
  query(['page']).custom((val) => { return mw.containsNoInvalidCharacters(val) }),
  query(['page']).optional().isInt({ min: CONSTANTS.PAGINATOR_PAGE }),
  // parseError,
  parseGetParams,
  controller.ALL_ORGS
)

router.get('/registryOrg/:identifier',
  /*
  #swagger.tags = ['Registry Organization']
  #swagger.operationId = 'getSingleRegistryOrg'
  #swagger.summary = "Retrieves information about a specific registry organization"
  #swagger.description = "
        <h2>Access Control</h2>
        <p>All authenticated users can access this endpoint</p>
        <h2>Expected Behavior</h2>
        <p><b>All Users:</b> Retrieves information about the specified registry organization</p>
  #swagger.parameters['identifier'] = {
    in: 'path',
    description: 'The identifier of the registry organization',
    required: true,
    type: 'string'
  }
  #swagger.parameters['$ref'] = [
    '#/components/parameters/apiEntityHeader',
    '#/components/parameters/apiUserHeader',
    '#/components/parameters/apiSecretHeader'
  ]
  #swagger.responses[200] = {
    description: 'The requested registry organization information is returned',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/org/get-org-response.json' }
      }
    }
  }
  #swagger.responses[400] = {
    description: 'Bad Request',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/bad-request.json' }
      }
    }
  }
  #swagger.responses[401] = {
    description: 'Not Authenticated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[404] = {
    description: 'Not Found',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[500] = {
    description: 'Internal Server Error',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  */
  mw.validateUser(true),
  param(['identifier']).isString().trim(),
  // parseError,
  parseGetParams,
  controller.SINGLE_ORG
)

router.post('/registryOrg',
  /*
  #swagger.tags = ['Registry Organization']
  #swagger.operationId = 'createRegistryOrg'
  #swagger.summary = "Creates a new registry organization (accessible to Secretariat only)"
  #swagger.description = "
        <h2>Access Control</h2>
        <p>Only users with Secretariat role can access this endpoint</p>
        <h2>Expected Behavior</h2>
        <p><b>Secretariat:</b> Creates a new registry organization</p>
  #swagger.parameters['$ref'] = [
    '#/components/parameters/apiEntityHeader',
    '#/components/parameters/apiUserHeader',
    '#/components/parameters/apiSecretHeader'
  ]
  #swagger.requestBody = {
    required: true,
    content: {
      'application/json': {
        schema: { $ref: '#/components/schemas/CreateOrgPayload' }
      }
    }
  }
  #swagger.responses[201] = {
    description: 'The registry organization was successfully created',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/org/create-org-response.json' }
      }
    }
  }
  #swagger.responses[400] = {
    description: 'Bad Request',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/bad-request.json' }
      }
    }
  }
  #swagger.responses[401] = {
    description: 'Not Authenticated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[403] = {
    description: 'Forbidden',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[500] = {
    description: 'Internal Server Error',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  */
  mw.validateUser(true),
  mw.onlySecretariat(true),
  body(['short_name']).isString().trim().notEmpty().isLength({ min: CONSTANTS.MIN_SHORTNAME_LENGTH, max: CONSTANTS.MAX_SHORTNAME_LENGTH }),
  body(['long_name']).isString().trim().notEmpty(),
  body(['authority.active_roles']).optional()
    .custom(isFlatStringArray)
    .customSanitizer(toUpperCaseArray)
    .custom(isOrgRole),
  body(['soft_quota']).optional().not().isArray().isInt({ min: CONSTANTS.MONGOOSE_VALIDATION.Org_policies_id_quota_min, max: CONSTANTS.MONGOOSE_VALIDATION.Org_policies_id_quota_max }).withMessage(errorMsgs.ID_QUOTA),
  body(['hard_quota']).optional().not().isArray().isInt({ min: CONSTANTS.MONGOOSE_VALIDATION.Org_policies_id_quota_min, max: CONSTANTS.MONGOOSE_VALIDATION.Org_policies_id_quota_max }).withMessage(errorMsgs.ID_QUOTA),
  // TODO: more validation needed?
  // parseError,
  parsePostParams,
  controller.CREATE_ORG
)

router.put('/registryOrg/:shortname',
  /*
  #swagger.tags = ['Registry Organization']
  #swagger.operationId = 'updateRegistryOrg'
  #swagger.summary = "Updates an existing registry organization (accessible to Secretariat only)"
  #swagger.description = "
        <h2>Access Control</h2>
        <p>Only users with Secretariat role can access this endpoint</p>
        <h2>Expected Behavior</h2>
        <p><b>Secretariat:</b> Updates an existing registry organization</p>
  #swagger.parameters['shortname'] = {
    in: 'path',
    description: 'The Shortname of the registry organization to update',
    required: true,
    type: 'string'
  }
  #swagger.parameters['$ref'] = [
    '#/components/parameters/apiEntityHeader',
    '#/components/parameters/apiUserHeader',
    '#/components/parameters/apiSecretHeader'
  ]
  #swagger.requestBody = {
    required: true,
    content: {
      'application/json': {
        schema: { $ref: '#/components/schemas/UpdateOrgPayload' }
      }
    }
  }
  #swagger.responses[200] = {
    description: 'The registry organization was successfully updated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/org/update-org-response.json' }
      }
    }
  }
  #swagger.responses[400] = {
    description: 'Bad Request',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/bad-request.json' }
      }
    }
  }
  #swagger.responses[401] = {
    description: 'Not Authenticated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[403] = {
    description: 'Forbidden',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[404] = {
    description: 'Not Found',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[500] = {
    description: 'Internal Server Error',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  */
  mw.validateUser(true),
  mw.onlySecretariat(true),
  param(['shortname']).isString().trim(),
  // TODO: do more validation here
  // parseError,
  parsePostParams,
  parseGetParams,
  controller.UPDATE_ORG
)

router.delete('/registryOrg/:identifier',
  /*
  #swagger.tags = ['Registry Organization']
  #swagger.operationId = 'deleteRegistryOrg'
  #swagger.summary = "Deletes an existing registry organization (accessible to Secretariat only)"
  #swagger.description = "
        <h2>Access Control</h2>
        <p>Only users with Secretariat role can access this endpoint</p>
        <h2>Expected Behavior</h2>
        <p><b>Secretariat:</b> Deletes an existing registry organization</p>
  #swagger.parameters['identifier'] = {
    in: 'path',
    description: 'The identifier of the registry organization to delete',
    required: true,
    type: 'string'
  }
  #swagger.parameters['$ref'] = [
    '#/components/parameters/apiEntityHeader',
    '#/components/parameters/apiUserHeader',
    '#/components/parameters/apiSecretHeader'
  ]
  #swagger.responses[200] = {
    description: 'The registry organization was successfully deleted',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/org/delete-org-response.json' }
      }
    }
  }
  #swagger.responses[400] = {
    description: 'Bad Request',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/bad-request.json' }
      }
    }
  }
  #swagger.responses[401] = {
    description: 'Not Authenticated',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[403] = {
    description: 'Forbidden',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  #swagger.responses[404] = {
    description: 'Not Found',
    content: {
      "application/json": {
        schema: { $ref: '../schemas/errors/generic.json' }
      }
    }
  }
  */
  mw.validateUser,
  // TODO: permissions
  param(['identifier']).isString().trim(),
  // parseError,
  parseDeleteParams,
  controller.DELETE_ORG
)

module.exports = router